pyproject.toml
[tool.poetry]
name = "maru"
version = "0.1.2"
description = "Morphological Analyzer for Russian 💬"
license = "MIT"
authors = ["Vladislav Blinov <cunningplan@yandex.ru>"]
repository = "https://github.com/chomechome/maru"
readme = "README.rst"
classifiers = [
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: Implementation :: CPython",
]

[tool.poetry.dependencies]
python = "^3.6"
numpy = "^1.15.0"
pymorphy2 = { version = "^0.8", extras = [ "fast" ] }
scipy = "^1.1.0"
keras = "^2.2.2"
tensorflow = ">=1.9.0, <1.15.0"
scikit-learn = "^0.19.0"
python-crfsuite = "^0.9.5"
lru-dict = "^1.1.6"
tensorflow-gpu = { version = ">=1.9.0, <1.15.0", optional = true }

[tool.poetry.dev-dependencies]
pytest = "^5.2.2"
pytest-cov = "^2.8.1"
pytest-xdist = "^1.30.0"
mypy = "^0.740"
flake8 = "^3.7.9"
flake8-isort = "^2.7.0"
flake8-builtins = "^1.4.1"
flake8-comprehensions = "^3.0.1"
flake8-debugger = "^3.2.1"
flake8-eradicate = "^0.2.3"
flake8-pytest = "^1.3"
pep8-naming = "^0.9.0"
hypothesis = "^4.44.2"
black = "^19.10b0"
isort = "^4.3.21"
unify = "^0.5"
pylint = "^2.4.3"
codecov = "^2.0.15"

[tool.poetry.extras]
gpu = ["tensorflow-gpu"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
